recyview

+++ mediapicker/src/main/java/com/dmcbig/mediapicker/PickerActivity.java
@@ -70,7 +70,7 @@ public class PickerActivity extends Activity implements DataCallback ,View.OnCli
         ArrayList<Media> medias =new ArrayList<>();
         ArrayList<Media> select=argsIntent.getParcelableArrayListExtra(PickerConfig.DEFAULT_SELECTED_LIST);
         int maxSelect=argsIntent.getIntExtra(PickerConfig.MAX_SELECT_COUNT,PickerConfig.DEFAULT_SELECTED_MAX_COUNT);
-        int maxSize=argsIntent.getIntExtra(PickerConfig.MAX_SELECT_SIZE,PickerConfig.DEFAULT_SELECTED_MAX_SIZE);
+        long maxSize=argsIntent.getLongExtra(PickerConfig.MAX_SELECT_SIZE,PickerConfig.DEFAULT_SELECTED_MAX_SIZE);
         gridAdapter =new  MediaGridAdapter(medias,this,select,maxSelect,maxSize);
         recyclerView.setAdapter(gridAdapter);
     }
